refactor loop closure into focused modules
